requirements.txt
duckduckgo-search>=5.0.0
git+https://github.com/randerzander/pyreadability.git@aed630c5a345aaf7432b16b7a45da3e71d5ae275
requests>=2.31.0
discord.py>=2.3.0
colorama>=0.4.6
pyyaml>=6.0.0
watchdog>=3.0.0
openai>=1.0.0
yt-dlp>=2023.0.0
pywhispercpp>=1.0.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.0
Wikipedia-API>=0.7.0
ffmpeg-python
lancedb>=0.26.0
sentencepiece>=0.2.0
